Support --disable-maintainer-mode
Adding AM_MAINTAINER_MODE([enable]) preserves the current behaviour of
rebuilding the autotools support by default if anything changed to make
it out of date -- which is desirable for people building from git to
avoid "Why is the build broken? Oh I forgot to rerun autogen.sh", which
is an all too frequent FAQ.

But it also allows it to be disabled, which is mostly useful for things
like distro package builds where the files should be preserved in their
original form, and configure shouldn't be run twice on the first build
if the timestamps were not strictly preserved for these files.

Signed-off-by: Ralph Giles <giles@mozilla.com>