mingw: Add a Dockerfile describing cross build.
For some reason, when I cross-compile opusfile on Fedora 23,
I get example binaries with exit immediately without error.

Building them on Fedora 20 (in a docker container) produced
working binaries. Until we figure out what the bug is, include
this for convenience.

The build doesn't actually work because of the 'make check'
steps in the Makefile, which require wine.