Move win32 package build details to mingw/README.md
In theory `mingw/Makefile` can build a binary release, but
it doensn't quite works and omits some niceties, so we still
need the detailed build instructions to compare with.

However, the details aren't relevent to the general release
checklist, so pretend `make -C ming package` works at that
level.

Also convert the mingw README to markdown for better
display on hosting sites.