mingw: Build against openssl 1.0.2u.
Download the most recent working release of openssl.
- Enable asm, since it seems to work now.
- Link ssl and crypto libraries statically, since the
  dll target seems to be broken.

The 1.0.2 series of openssl releases is no longer supported.
However, 1.1.1 and later broke the hooks we were using to
reference the Windows certificate store, so this is also
the newest release where opusurl works.

As such, this shouldn't be used in production (unless you have a
support contract for openssl!) but it's a useful reference
configuration to have in the repository.

Hopefully a replacement api will be available in openssl 3 or 4
and we can port to that for future releases.