cmake: add -lsocket if necessary
On OpenIndiana (Illumos) the cmake build fails with indirect
symbol references to functions declared by `sys/socket.h`.

To address this, check for `getsockopt` linkage, and if it
fails try again with `-lsocket` which provides the symbols
on that operating system.

This doesn't seem to be necessary with the autotools build;
perhaps libtool is passing something extra to the linker
to have it accept the transitive dependencies.

Fixes https://github.com/xiph/opusfile/issues/60
Fixes https://gitlab.xiph.org/xiph/opusfile/-/issues/2337