gitlab-ci: Add default cmake and cmake-nohttp jobs
In addition to testing the build with a legacy cmake-3.16 upstream
package, install the runner's default cmake version and test against
that for broader coverage. On the current `gcc:9` container image
this is cmake 3.18.4.

Also test build with `-D OP_DISABLE_HTTP=YES` to confirm the build
completes with this option, if not that it does what it's supposed
to. This should protect against regression of the earlier fix.